\name{plgp-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
\alias{calc.ecis}
\alias{calc.eis}
\alias{calc.ents}
\alias{calc.ieci}
\alias{calc.iecis}
\alias{calc.alcs}
\alias{calc.vars}
\alias{calc.ktKik.x}
\alias{calc2.ktKik.x}
\alias{covar}
\alias{covar.sep}
\alias{covar.sim}
\alias{cv.folds}
\alias{dist2covar.symm}
\alias{distance}
\alias{EI}
\alias{ei.adapt}
\alias{entropy}
\alias{entropy.adapt}
\alias{entropy.bvsb}
\alias{findmin.ConstGP}
\alias{findmin.GP}
\alias{hist.particle.params}
\alias{ieci.adapt}
\alias{alc.adapt}
\alias{mindist.adapt}
\alias{var.adapt}
\alias{ieci.const.adapt}
\alias{alc.const.adapt}
\alias{ieci.ConstGP}
\alias{alc.ConstGP}
\alias{ieci.GP}
\alias{alc.GP}
\alias{lpost.GP}
\alias{PL.clear}
\alias{pred.mean.GP}
\alias{resample}
\alias{tquants}
\alias{unif.propose.pos}
\alias{mvnorm.propose.rw}
\alias{updat.GP}
\alias{util.GP}
\alias{renorm.weights}
\alias{renorm.lweights}
\alias{latents.CGP}
\alias{phist}
\alias{getmap.CGP} 
\alias{getmap.GP}

\title{Internal plgp Functions}
\description{
  Internal \pkg{plgp} functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :))
}
\keyword{ internal }
